Write out the cdash PR url.
This is very useful for people without jenkins/cern accounts.

This patch filters out the irrelevant cdash information and shows only the ones
related to the current PR.

This is suboptimal because it will not connect the concrete build id with the
concrete uploaded artifact in cdash but it is better than nothing.

CMake 3.15 has an option BUILD_ID to its ctest_submit which we can use to output
a concrete link to a concrete build subjob.